Update attribute encoding, minor UI / pipeline fixes
